Fix warning when using Mocking and precompilation
Previously using Mocking within a package and then pre-compiling that
package would result in the warning:

    WARNING: eval from module Mocking to ...:
    Expr(:module, true, :##285, Expr(:block, Expr(:line, 108, :~/.julia/v0.5/Mocking/src/Mocking.jl)::Any, Expr(:line, 108, :~/.julia/v0.5/Mocking/src/Mocking.jl)::Any)::Any)::Any
      ** incremental compilation may be broken for this module **

The warning was caused from the line:

    m = Core.eval(Mocking, :(module $(gensym()) end))

which creates a new module on-the-fly. Since Mocking.jl only works
with testing when --compilecache=no (not using precompilation) Mocking
should always be safe to use with pre-compiled modules.

Ultimately, the fix was to avoid creating `PatchEnv` during compilation.