Remove implicit enabling of Mocking
In Julia 0.5- we could detect when `Pkg.test` was executed by checking
if `basename(Base.PROGRAM_FILE) == "runtests.jl"` and then automatically
enabling Mocking. At first appeared to work as we intended but later it
was found that packages that did not require Mocking but did require a
package that did would cause Mocking to be enabled. Since Mocking can
warn when it may not work properly this could lead to unnecessary
confusion.

The solution at the moment is to only allow explicit enabling of
Mocking. In the future we may be able to restore the implicit enabling
if we can meet the following conditions:

1. A test suite is being run
2. The test suite itself has imported Mocking