Changes required for the Julia PR #17057
The Julia PR #17057 (the fix for issue #265) introduced some changes
which made it invalid to call a function which was compiled after the
call was compiled. This particular problem would occur when using the
`apply` do-block.

Note we could mostly avoid using `eval(Expr(:call, ...` if we force
users to make a PatchEnv prior to using the apply do-block.
Unfortunately this solution does not work with mocking a nested function
and also is less user friendly.