[#108] Support mocks running a background task
Allow patching a mocked call that is running in an older world age by changing the macro to invoke the patch through `invokelatest`.

The call was already inherently going to be a dynamic dispatch anyway, since by definition the whole point is that we want to substitute in a new function later. Invokelatest will just allow us to substitute in a function that was even _defined_ later, after this function had started running.

This doesn't add any extra overhead, nor introduce any restrictions. The only thing that it does is increase the scenarios that Mocking supports.