Fix up the scoping for the `@mock` macro.
Previously, the macro required `Mocking` to be visible in the caller's context:
It was invoking `Mocking.activated()`, for exmaple.  Now it invokes
`$activated()`, which makes sure that the code uses the right version of
`activated`.  Similarly for `get_alternate` and `Patch`.

Also, the macro was doing some unceessary `gensym` calls, which macro hygene can
take care of automatically.